fixing styles and list
