+++ packages/app/src/assets/poolTypes/index.ts
@@ -1,0 +1,3 @@
+export { default as DirectPayments } from './directPayments.png';
+export { default as Segmented } from './segmented.png';
+export { default as Community } from './community.png';
