+++ packages/app/src/custom.d.ts
@@ -6,8 +6,4 @@ declare module '*.png' {
   const content: any;
   export default content;
 }
-declare module '*.jpg' {
-  const content: any;
-  export default content;
-}
 //
